+++ apps/client/src/components/AuthHandler/AuthHandler.tsx
@@ -4,17 +4,18 @@ import {
 } from "../pages/SignIn/hooks/useUserInfoQuery/useUserInfoQuery";
 import {Progress} from "@mantine/core";
 import {Section} from "../Section/Section";
-import {SomethingWentWrong} from "../pages/Tasks/Tasks";
 import {
 	useIncrementOnInterval
 } from "../../hooks/useIncrementOnInterval/useIncrementOnInterval";
+import {SomethingWentWrong} from "../SomethingWentWrong/SomethingWentWrong";
+import {IChildren} from "../../interfaces";
 
 /**
  * @description Used as a way of rendering errors, loadings, and calling the useUserInfoQuery prior to the Router component.
  * @param children
  * @constructor
  */
-export const AuthHandler: FunctionComponent<{ children: JSX.Element }> = ({children}) => {
+export const AuthHandler: FunctionComponent<IChildren> = ({children}) => {
 	// If getting the auth state takes unexpectedly long, we'll show a loading indicator.
 	// The rest of the app does not load until the auth state is ready, otherwise the user would get a flash of the wrong layout and routes.
 	const {isLoading, isError, error} = useUserInfoQuery();
@@ -42,5 +43,5 @@ export const AuthHandler: FunctionComponent<{ children: JSX.Element }> = ({child
 		);
 	}
 
-	return children;
+	return <>{children}</>;
 }
