+++ apps/client/src/components/DnDReactTable/THead/Search/interfaces.ts
@@ -1,6 +1,6 @@
 import {BaseArray} from "@prioritea/types";
 
-export interface SearchBarProps {
+export interface SearchProps {
 	visibleColumnsLength: number;
 	preGlobalFilteredRows: BaseArray;
 	globalFilter: string | undefined;
