+++ apps/client/src/components/NavLink/NavLink.tsx
@@ -1,86 +1,81 @@
-import { createStyles } from "@mantine/core";
-import { FunctionComponent, MouseEventHandler, ReactNode } from "react";
-import { NavLink as RouterNavLink } from "react-router-dom";
-
-interface LinkProps {
-  end?: boolean;
-  to: string;
-  onClick?: MouseEventHandler<HTMLAnchorElement>;
-  children: ReactNode;
-}
+import {createStyles} from "@mantine/core";
+import {FunctionComponent} from "react";
+import {NavLink as RouterNavLink} from "react-router-dom";
+import {LinkProps} from "./interfaces";
 
 export const NavLink: FunctionComponent<LinkProps> = ({
-  end,
-  to,
-  children,
-  onClick,
-}) => {
-  const useStyles = createStyles((theme, _params, getRef) => {
-        const icon = getRef("icon");
+														  end,
+														  to,
+														  children,
+														  onClick,
+													  }) => {
+	const useStyles = createStyles((theme, _params, getRef) => {
+		const icon = getRef("icon");
 
-    return {
-    link: {
-      textTransform: "capitalize",
-      display: "flex",
-      alignItems: "center",
-      lineHeight: 1,
-      padding: "8px 12px",
-      borderRadius: theme.radius.sm,
-      textDecoration: "none",
-      color:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[0]
-          : theme.colors.gray[7],
-      fontSize: theme.fontSizes.sm,
-      fontWeight: 500,
+		return {
+			link: {
+				textTransform: "capitalize",
+				display: "flex",
+				alignItems: "center",
+				lineHeight: 1,
+				padding: "8px 12px",
+				borderRadius: theme.radius.sm,
+				textDecoration: "none",
+				color:
+					theme.colorScheme === "dark"
+						? theme.colors.dark[0]
+						: theme.colors.gray[7],
+				fontSize: theme.fontSizes.sm,
+				fontWeight: 500,
 
-      "&:hover": {
-        backgroundColor:
-          theme.colorScheme === "dark"
-            ? theme.colors.dark[6]
-            : theme.colors.gray[0],
-      },
+				"&:hover": {
+					backgroundColor:
+						theme.colorScheme === "dark"
+							? theme.colors.dark[6]
+							: theme.colors.gray[0],
+				},
 
-      [theme.fn.smallerThan("sm")]: {
-        borderRadius: 0,
-        padding: theme.spacing.md,
-      },
-    },
+				[theme.fn.smallerThan("sm")]: {
+					borderRadius: 0,
+					padding: theme.spacing.md,
+				},
+			},
 
-    linkActive: {
-        "&, &:hover": {
-          backgroundColor:
-            theme.colorScheme === "dark"
-            // @ts-ignore
-              ? theme.fn.rgba(theme.colors[theme.primaryColor][8], 0.25)
-              // @ts-ignore
-              : theme.colors[theme.primaryColor][0],
-          color:
-            theme.colorScheme === "dark"
-              ? theme.white
-              // @ts-ignore
-              : theme.colors[theme.primaryColor][7],
-          [`& .${icon}`]: {
-            color:
-            // @ts-ignore
-            theme.colors[
-              theme.primaryColor][theme.colorScheme === "dark" ? 5 : 7],
-          },
-        },
-      },
-  }});
-  const { classes, cx } = useStyles();
+			linkActive: {
+				"&, &:hover": {
+					backgroundColor:
+						theme.colorScheme === "dark"
+							// @ts-ignore
+							? theme.fn.rgba(theme.colors[theme.primaryColor][8], 0.25)
+							// @ts-ignore
+							: theme.colors[theme.primaryColor][0],
+					color:
+						theme.colorScheme === "dark"
+							? theme.white
+							// @ts-ignore
+							: theme.colors[theme.primaryColor][7],
+					[`& .${icon}`]: {
+						color:
+						// @ts-ignore
+							theme.colors[
+								theme.primaryColor][theme.colorScheme === "dark" ? 5 : 7],
+					},
+				},
+			},
+		}
+	});
+	const {classes, cx} = useStyles();
 
-  return (
-    <RouterNavLink
-      className={({ isActive }) =>
-        cx(classes.link, { [classes.linkActive]: isActive })
-      }
-      end={end}
-      to={to}
-      onClick={onClick}
-    >
-      {children}
-    </RouterNavLink>
-  );
+	return (
+		<RouterNavLink
+			className={({isActive}) =>
+				cx(classes.link, {[classes.linkActive]: isActive})
+			}
+			end={end}
+			to={to}
+			onClick={onClick}
+		>
+			{children}
+		</RouterNavLink>
+	);
 };
