+++ apps/client/src/components/NavLink/interfaces.ts
@@ -1,0 +1,8 @@
+import {MouseEventHandler} from "react";
+import {IChildren} from "../../interfaces";
+
+export interface LinkProps extends IChildren {
+	end?: boolean;
+	to: string;
+	onClick?: MouseEventHandler<HTMLAnchorElement>;
+}
