--- apps/client/src/components/Providers/types.ts
@@ -1,5 +1,0 @@
-import {ReactNode} from "react";
-
-export interface ProvidersProps {
-	children: ReactNode;
-}
