+++ apps/client/src/components/Router/protectedRoutes.ts
@@ -1,3 +1,6 @@
+/**
+ * Routes that require authentication.
+ */
 export const protectedRoutes = [
 	"/",
 	"/account"
