--- apps/client/src/components/pages/Tasks/CreateTaskModal/interfaces.d.ts
@@ -1,4 +1,0 @@
-export interface CreateTaskModalProps {
-	isOpen: boolean;
-	onClose: () => void;
-}
