+++ apps/client/src/components/pages/Tasks/DeleteTasks/interfaces.ts
@@ -1,0 +1,4 @@
+export interface DeleteTasksProps {
+	disabled: boolean;
+	selectedTasksIds: Array<string>;
+}
