+++ apps/client/src/components/pages/Tasks/UpdateTaskModal/UpdateTaskModal.tsx
@@ -5,42 +5,36 @@ import {
 import {Controller, SubmitHandler, useForm} from "react-hook-form";
 import {zodResolver} from "@hookform/resolvers/zod";
 import {
+	ActionIcon,
 	Button,
 	Group,
 	Modal,
 	NumberInput,
 	Select,
-	TextInput
+	TextInput,
+	Tooltip
 } from "@mantine/core";
 import {FieldError} from "../../../FieldError/FieldError";
 import {UpdateTaskModalProps} from "./interfaces";
-import {UpdateTaskDto} from "@prioritea/types";
+import {Priority, Status, UpdateTaskDto} from "@prioritea/types";
 import {updateTaskSchema} from "@prioritea/validation";
 import {useTasksQuery} from "../hooks/useTasksQuery/useTasksQuery";
-import {formatTaskStatus} from "../Tasks";
-
-export enum Status {
-	IDLE = "IDLE",
-	IN_PROGRESS = "IN_PROGRESS",
-	COMPLETED = "COMPLETED",
-}
-
-export enum Priority {
-	ONE = 1,
-	TWO = 2,
-	THREE = 3,
-	FOUR = 4,
-	FIVE = 5,
-	MIN = Priority.ONE,
-	MAX = Priority.FIVE,
-}
-
+import {formatTaskStatus} from "../utils/format-task-status/format-task-status";
+import {Pencil} from "tabler-icons-react";
+import {useToggle} from "../../../../hooks/useToggle/useToggle";
 
 export const UpdateTaskModal: FunctionComponent<UpdateTaskModalProps> = ({
-							id,												 isOpen,
-																			 onClose
+																			 id,
+																			 disabled
+
 																		 }) => {
-	const {mutateAsync} = useUpdateTaskMutation(onClose);
+	const [
+		modalIsOpen,
+		,
+		openModal,
+		closeModal
+	] = useToggle();
+	const {mutateAsync} = useUpdateTaskMutation(closeModal);
 	const {data: tasks} = useTasksQuery();
 	const task = useMemo(() => tasks?.find((task) => task.id === id), [tasks?.length, id]);
 	const {
@@ -58,15 +52,15 @@ export const UpdateTaskModal: FunctionComponent<UpdateTaskModalProps> = ({
 		resolver: zodResolver(updateTaskSchema)
 	});
 	const onSubmit: SubmitHandler<Omit<UpdateTaskDto, 'id'>> = async ({
-														  priority,
-															  description,
-																		status
-														  }) => {
+																		  priority,
+																		  description,
+																		  status
+																	  }) => {
 		if (!id) return;
 
 		await mutateAsync({id, priority, description, status});
 
-		onClose();
+		closeModal();
 	};
 
 	useEffect(() => {
@@ -75,57 +69,69 @@ export const UpdateTaskModal: FunctionComponent<UpdateTaskModalProps> = ({
 
 
 	return (
-		<Modal
-			opened={isOpen}
-			onClose={onClose}
-			title="Update a task"
-		>
-			<form noValidate onSubmit={handleSubmit(onSubmit)}>
-				<Group direction="column" grow>
-					<Controller
-						control={control}
-						name={"priority"}
-						render={({field}) => (
-							<NumberInput
-								label="Priority"
-								placeholder="Type here.."
-								min={Priority.MIN}
-								max={Priority.MAX}
-								{...field}
-							/>
-						)}
-					/>
-					<FieldError field={errors.priority}/>
-					<TextInput
-						label="Description"
-						placeholder={task?.description}
-						{...register("description")}
-					/>
-					<FieldError field={errors.description}/>
-					<Controller
-						control={control}
-						name={"status"}
-						render={({field}) => (
-							<Select
-								label="Status"
-								placeholder="Pick one.."
-								data={Object.values(Status).map((status) => ({
-									value: status,
-									label: formatTaskStatus(status)
-								}))}
-								{...field}
-							/>
-						)}
-					/>
-					<FieldError field={errors.status}/>
-				</Group>
-				<Group position="apart" mt="xl">
-					<Button type="submit" style={{textTransform: "capitalize"}}
-							variant="filled">
-						Update
-					</Button>
-				</Group>
-			</form>
-		</Modal>
+		<>
+			<Modal
+				opened={modalIsOpen}
+				onClose={closeModal}
+				title="Update a task"
+			>
+				<form noValidate onSubmit={handleSubmit(onSubmit)}>
+					<Group direction="column" grow>
+						<Controller
+							control={control}
+							name={"priority"}
+							render={({field}) => (
+								<NumberInput
+									label="Priority"
+									placeholder="Type here.."
+									min={Priority.MIN}
+									max={Priority.MAX}
+									{...field}
+								/>
+							)}
+						/>
+						<FieldError field={errors.priority}/>
+						<TextInput
+							label="Description"
+							placeholder={task?.description}
+							{...register("description")}
+						/>
+						<FieldError field={errors.description}/>
+						<Controller
+							control={control}
+							name={"status"}
+							render={({field}) => (
+								<Select
+									label="Status"
+									placeholder="Pick one.."
+									data={Object.values(Status).map((status) => ({
+										value: status,
+										label: formatTaskStatus(status)
+									}))}
+									{...field}
+								/>
+							)}
+						/>
+						<FieldError field={errors.status}/>
+					</Group>
+					<Group position="apart" mt="xl">
+						<Button type="submit"
+								style={{textTransform: "capitalize"}}
+								variant="filled">
+							Update
+						</Button>
+					</Group>
+				</form>
+			</Modal>
+			<Tooltip label={'Update selected task'} withArrow>
+				<ActionIcon mb="1rem" size={24} color="primary" radius="xl"
+							variant="filled"
+							disabled={disabled}
+							onClick={openModal}
+				>
+					<Pencil size={18}/>
+				</ActionIcon>
+			</Tooltip>
+		</>
 	);
 }
