+++ apps/client/src/components/pages/Tasks/UpdateTaskModal/hooks/useUpdateTaskMutation/useUpdateTaskMutation.tsx
@@ -1,10 +1,13 @@
 import {useMutation, useQueryClient} from "react-query";
 import {TasksApi} from "../../../../../../api/tasks-api/tasks-api";
 import {Tasks, UpdateTaskDto} from "@prioritea/types";
-import {errorToast} from "../../../Tasks";
+import {
+	useErrorToast
+} from "../../../../../../hooks/useErrorToast/useErrorToast";
 
 export const useUpdateTaskMutation = (onClose: () => void) => {
 	const queryClient = useQueryClient();
+	const errorToast = useErrorToast();
 
 	return useMutation(async ({
 								  id,
@@ -21,7 +24,9 @@ export const useUpdateTaskMutation = (onClose: () => void) => {
 
 		return data.data.tasks;
 	}, {
+		// Optimistic updates
 		async onMutate(editedTask) {
+			// Make sure that meanwhile tasks don't get overridden
 			await queryClient.cancelQueries(['tasks']);
 
 			const prevTasks = queryClient.getQueryData(['tasks']) as Tasks;
@@ -29,6 +34,7 @@ export const useUpdateTaskMutation = (onClose: () => void) => {
 			queryClient.setQueryData(['tasks'], (prev) =>
 				// @ts-ignore
 				prev.map((task) => {
+					// Override the task with the edited one
 					if (task.id === editedTask.id) {
 						return editedTask;
 					}
@@ -36,11 +42,14 @@ export const useUpdateTaskMutation = (onClose: () => void) => {
 					return task;
 				}));
 
+			// Pass the previous tasks to the error handler
 			return {prevTasks};
 		},
+		// Revert optimistic updates on error
 		onError(err, _editedTask, context: { prevTasks: Tasks } | undefined) {
 			queryClient.setQueryData(['tasks'], context?.prevTasks);
 
+			// 404 is the most likely error when using the UI
 			// @ts-ignore
 			if (err.response?.status === 404) {
 				onClose();
@@ -50,8 +59,10 @@ export const useUpdateTaskMutation = (onClose: () => void) => {
 				return;
 			}
 
+			// Fallback error message
 			errorToast("Failed to update task");
 		},
+		// Regardless of the outcome, make sure the data is correct.
 		onSettled() {
 			queryClient.invalidateQueries(['tasks']);
 		}
