+++ apps/client/src/components/pages/Tasks/UpdateTaskModal/interfaces.d.ts
@@ -1,5 +1,4 @@
 export interface UpdateTaskModalProps {
 	id: string;
-	isOpen: boolean;
-	onClose: () => void;
+	disabled: boolean;
 }
