+++ apps/client/src/components/pages/Tasks/utils/format-task-status/format-task-status.ts
@@ -1,0 +1,3 @@
+import {stringUtils} from "@prioritea/utils";
+
+export const formatTaskStatus = (status: string) => stringUtils(status).toKebabCase().toCapitalized().string
