+++ apps/client/src/config.ts
@@ -1,2 +1,5 @@
+/**
+ * To be used with axios
+ */
 export const TIMEOUT_IN_MS = 5000;
