+++ apps/client/src/lib/axios-client.ts
@@ -1,8 +1,11 @@
 import axios from "axios";
-import { TIMEOUT_IN_MS } from "../config";
+import {TIMEOUT_IN_MS} from "../config";
 
+/**
+ * An instance of axios with the app's url, a timeout, and the auth header set to be sent with every request.
+ */
 export const axiosClient = axios.create({
-  baseURL: `${import.meta.env.VITE_API_URL}/`,
-  timeout: TIMEOUT_IN_MS,
-  withCredentials: true,
+	baseURL: `${import.meta.env.VITE_API_URL}/`,
+	timeout: TIMEOUT_IN_MS,
+	withCredentials: true,
 });
