+++ apps/client/src/lib/query-client.ts
@@ -1,4 +1,6 @@
 import {QueryClient} from "react-query";
 
-
+/**
+ * react-query's QueryClient
+ */
 export const queryClient = new QueryClient();
