--- apps/client/src/types.ts
@@ -1,4 +1,0 @@
-import { Dispatch, SetStateAction } from "react";
-
-export type SetState<TState> = Dispatch<SetStateAction<TState>>;
-
